# server/immich-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/search": {
      "get": {
        "operationId": "search",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "IMAGE",
                "VIDEO",
                "AUDIO",
                "OTHER"
              ],
              "type": "string"
            }
          },
          {
            "name": "isFavorite",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "exifInfo.city",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exifInfo.state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exifInfo.country",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exifInfo.make",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exifInfo.model",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "smartInfo.objects",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "smartInfo.tags",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "recent",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "motion",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Search"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/time-bucket/assets": {
      "get": {
        "operationId": "getByTimeBucket",
        "parameters": [
          {
            "name": "size",
            "required": true,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TimeBucketSize"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "isArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isFavorite",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "timeBucket",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Time Bucket"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/person/{id}/assets": {
      "get": {
        "operationId": "getPersonAssets",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Person"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/upload": {
      "post": {
        "operationId": "uploadFile",
        "parameters": [
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Asset Upload Information",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetFileUploadResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/download/{assetId}": {
      "get": {
        "operationId": "downloadFile",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/download-files": {
      "post": {
        "operationId": "downloadFiles",
        "parameters": [
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadFilesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "description": ""
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/download-library": {
      "get": {
        "operationId": "downloadLibrary",
        "description": "Current this is not used in any UI element",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/file/{assetId}": {
      "get": {
        "operationId": "serveFile",
        "parameters": [
          {
            "name": "isThumb",
            "required": false,
            "in": "query",
            "schema": {
              "title": "Is serve thumbnail (resize) file",
              "type": "boolean"
            }
          },
          {
            "name": "isWeb",
            "required": false,
            "in": "query",
            "schema": {
              "title": "Is request made from web",
              "type": "boolean"
            }
          },
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/thumbnail/{assetId}": {
      "get": {
        "operationId": "getAssetThumbnail",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ThumbnailFormat"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/curated-objects": {
      "get": {
        "operationId": "getCuratedObjects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CuratedObjectsResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/curated-locations": {
      "get": {
        "operationId": "getCuratedLocations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CuratedLocationsResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/search-terms": {
      "get": {
        "operationId": "getAssetSearchTerms",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/search": {
      "post": {
        "operationId": "searchAsset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchAssetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/count-by-user-id": {
      "get": {
        "operationId": "getAssetCountByUserId",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetCountByUserIdResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/stat/archive": {
      "get": {
        "operationId": "getArchivedAssetCountByUserId",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetCountByUserIdResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/asset/map-marker": {
      "get": {
        "operationId": "getMapMarkers",
        "description": "Get all assets that have GPS information embedded",
        "parameters": [
          {
            "name": "isFavorite",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MapMarkerResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/asset/{deviceId}": {
      "get": {
        "operationId": "getUserAssetsByDeviceId",
        "description": "Get all asset of a device that are in the database, ID only.",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Immich",
    "description": "Immich API",
    "version": "1.55.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "immich_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "SearchResponseDto": {
        "type": "object",
        "properties": {
          "albums": {
            "$ref": "#/components/schemas/SearchAlbumResponseDto"
          },
          "assets": {
            "$ref": "#/components/schemas/SearchAssetResponseDto"
          }
        },
        "required": [
          "albums",
          "assets"
        ]
      },
...
      "TimeBucketSize": {
        "type": "string",
        "enum": [
          "DAY",
          "MONTH"
        ]
      },
...
      "AssetFileUploadResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "duplicate": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "duplicate"
        ]
      },
      "DownloadFilesDto": {
        "type": "object",
        "properties": {
          "assetIds": {
            "title": "Array of asset ids to be downloaded",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "assetIds"
        ]
      },
      "ThumbnailFormat": {
        "type": "string",
        "enum": [
          "JPEG",
          "WEBP"
        ]
      },
...
      "SearchAssetDto": {
        "type": "object",
        "properties": {
          "searchTerm": {
            "type": "string"
          }
        },
        "required": [
          "searchTerm"
        ]
      },
...
    }
  }
}